requirements.txt
streamlit==1.40.1
pandas==2.2.3
python-dotenv==0.21.0
pymysql==1.0.2
sqlalchemy==2.0.34
bcrypt==3.2.0
openai==1.59.3
pytorch==2.5.1
nltk==3.9.1
scikit-learn==1.5.2
pinecone-client==5.0.1
sentence-transformers==3.3.1